Finished task F. Refer to README for details.

+++ src/main/java/com/example/demo/controllers/AddProductController.java
@@ -56,7 +56,7 @@ public class AddProductController {
             ProductService productService = context.getBean(ProductServiceImpl.class);
             Product product2 = new Product();
             try {
-                product2 = productService.findById((int) product.getId());
+                product2 = productService.findById((long) product.getId());
             } catch (Exception e) {
                 System.out.println("Error Message " + e.getMessage());
             }
@@ -75,7 +75,7 @@ public class AddProductController {
         else {
             ProductService repo = context.getBean(ProductServiceImpl.class);
             if(product.getId()!=0) {
-                Product product2 = repo.findById((int) product.getId());
+                Product product2 = repo.findById((long) product.getId());
                 PartService partService1 = context.getBean(PartServiceImpl.class);
                 if(product.getInv()- product2.getInv()>0) {
                     for (Part p : product2.getParts()) {
@@ -97,7 +97,7 @@ public class AddProductController {
     public String showProductFormForUpdate(@RequestParam("productID") int theId, Model theModel) {
         theModel.addAttribute("parts", partService.findAll());
         ProductService repo = context.getBean(ProductServiceImpl.class);
-        Product theProduct = repo.findById(theId);
+        Product theProduct = repo.findById((long) theId);
         product1=theProduct;
     //    this.product=product;
         //set the employ as a model attibute to prepopulate the form
@@ -115,7 +115,7 @@ public class AddProductController {
     @GetMapping("/deleteproduct")
     public String deleteProduct(@RequestParam("productID") int theId, Model theModel) {
         ProductService productService = context.getBean(ProductServiceImpl.class);
-        Product product2=productService.findById(theId);
+        Product product2=productService.findById((long) theId);
         for(Part part:product2.getParts()){
             part.getProducts().remove(product2);
             partService.save(part);
