+++ src/main/java/com/example/demo/validators/EnufPartsValidator.java
@@ -31,7 +31,7 @@ public class EnufPartsValidator implements ConstraintValidator<ValidEnufParts, P
         if(context!=null)myContext=context;
         ProductService repo = myContext.getBean(ProductServiceImpl.class);
         if (product.getId() != 0) {
-            Product myProduct = repo.findById((int) product.getId());
+            Product myProduct = repo.findById((long) product.getId());
             for (Part p : myProduct.getParts()) {
                 if (p.getInv()<(product.getInv()-myProduct.getInv()))return false;
             }
